backend/requirements.txt
# Projeto Nix - Python Dependencies

# Django
Django>=5.0,<5.1
psycopg2-binary>=2.9.9  # PostgreSQL adapter

# Django Rest Framework
djangorestframework>=3.14.0
drf-spectacular>=0.27.0  # API Documentation (Swagger/ReDoc)
django-filter>=23.5  # Filtering for DRF
djangorestframework-simplejwt>=5.3.0  # JWT Authentication

# NFe XML parsing
lxml>=4.9.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.20.0

# CORS
django-cors-headers>=4.3.1

# Image processing (for logo, fotos)
Pillow>=10.1.0

# Environment variables
python-dotenv>=1.0.0

# Development only (optional)
django-debug-toolbar>=4.2.0  # Debug panel
ipython>=8.18.1  # Better shell
